Shifting functionality to service, currently not compiling
